fix: use isomorphic layout effect for mobile hook
