+++ src/hooks/use-mobile.tsx
@@ -1,5 +1,8 @@
 import * as React from "react"
 
+const useIsoLayoutEffect =
+  typeof window !== "undefined" ? React.useLayoutEffect : React.useEffect
+
 /**
  * Screen width (in pixels) below which layout is considered mobile.
  * Exported for reuse to ensure consistent breakpoint behavior.
@@ -9,7 +12,7 @@ export const MOBILE_BREAKPOINT = 768
 export function useIsMobile() {
   const [isMobile, setIsMobile] = React.useState(false)
 
-  React.useEffect(() => {
+  useIsoLayoutEffect(() => {
     if (typeof window === "undefined") return
     const mql = window.matchMedia(`(max-width: ${MOBILE_BREAKPOINT - 1}px)`)
     const onChange = () => {
